requirements.txt
fastapi
uvicorn
python-dotenv
youtube-transcript-api
requests
google-genai
yt_dlp